Support multiple versions of spark
