+++ spark/spark-3.1/src/main/scala/org/sedona/sql/parser/SedonaSqlAstBuilder.scala
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.spark.sql.parser
+package org.apache.sedona.sql.parser
 
 import org.apache.spark.sql.catalyst.parser.SqlBaseParser._
 import org.apache.spark.sql.execution.SparkSqlAstBuilder
